Abstract out the validation used for CSV
